requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0
astroid==2.11.3; python_full_version >= '3.6.2'
cachetools==5.0.0; python_version ~= '3.7'
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.4.1
dill==0.3.4; python_version >= '2.7' and python_version != '3.0'
django-ckeditor==6.4.0
django-filer==2.2
django-js-asset==1.2.2
django-mptt==0.13.4
django-phonenumber-field[phonenumberslite]==6.1.0
django-polymorphic==3.0.0
django==3.2.12
djangocms-admin-style==3.0.0
djangorestframework==3.13.1
drf-yasg==1.20.0
easy-thumbnails==2.8.1
google-api-core==2.7.2; python_version >= '3.6'
google-api-python-client==2.45.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-auth==2.6.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos==1.56.0; python_version >= '3.6'
hijri-converter==2.2.3
httplib2==0.20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3
inflection==0.5.1; python_full_version >= '3.5.0'
isort==5.10.1; python_version < '4.0' and python_full_version >= '3.6.1'
itypes==1.2.0
jinja2==3.1.1; python_version >= '3.7'
lazy-object-proxy==1.7.1; python_version >= '3.6'
lxml==4.7.1
markupsafe==2.1.1; python_version >= '3.7'
mccabe==0.7.0; python_version >= '3.6'
oauthlib==3.2.0; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
phonenumberslite==8.12.46
pillow==9.0.1
platformdirs==2.5.2; python_version >= '3.7'
protobuf==3.20.1rc1; python_version >= '3.7'
psycopg2-binary
pyasn1-modules==0.2.8
pyasn1==0.4.8
pylint-django==2.5.3
pylint-plugin-utils==0.7; python_full_version >= '3.6.2'
pylint==2.13.7; python_full_version >= '3.6.2'
pyparsing==3.0.8; python_version >= '3.1'
pypdf2==1.27.7
pytz==2021.3
python-gettext==5.0 
reportlab==3.6.6
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.27.1
rsa==4.8; python_version >= '3.6'
ruamel.yaml.clib==0.2.6; python_version < '3.11' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.21; python_version >= '3'
setuptools==62.1.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.2.0; python_version < '3.10'
unidecode==1.1.2
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.8
webencodings==0.5.1
wrapt==1.14.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

